[Marketing] Enable caching dependency archives
Increases security by preventing changes in remote code.

Metrics:

  * Build time drops from 15s to 10s.
  * Adds 28 MB of compressed size to the git database

+++ deploy.sh
@@ -200,8 +200,12 @@ PHPMYADMIN_URL="https://files.phpmyadmin.net/phpMyAdmin/${PHPMYADMIN_VERSION}/ph
 # TODO: Retry several times if a 404 is given (wget exits with code 8)
 # A 404 occurred once with wordpress.com
 echo "Downloading dependencies..."
-wget --no-show-progress "${WORDPRESS_URL}" -O "${DOWNLOAD_DIR}/wordpress-${WORDPRESS_VERSION}.zip"
-$EXCLUDE_PHPMYADMIN || wget --no-show-progress "${PHPMYADMIN_URL}" -O "${DOWNLOAD_DIR}/phpmyadmin-${PHPMYADMIN_VERSION}.zip"
+if [ ! -f "${DOWNLOAD_DIR}/wordpress-${WORDPRESS_VERSION}.zip" ]; then
+  wget --no-show-progress "${WORDPRESS_URL}" -O "${DOWNLOAD_DIR}/wordpress-${WORDPRESS_VERSION}.zip"
+fi
+if [ ! -f "${DOWNLOAD_DIR}/phpmyadmin-${PHPMYADMIN_VERSION}.zip" ]; then
+  $EXCLUDE_PHPMYADMIN || wget --no-show-progress "${PHPMYADMIN_URL}" -O "${DOWNLOAD_DIR}/phpmyadmin-${PHPMYADMIN_VERSION}.zip"
+fi
 
 # Extract dependencies
 echo "Extracting dependencies..."
